fixed #75, #83, #84

+++ deploy_osx.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
+CLEANCACHE="$1"
 
 launchctl unload ~/Library/LaunchAgents/com.homebridge.server.plist
-rm -f ~/.homebridge/accessories/cachedAccessories
+if  [ "" != "$CLEANCACHE" ]
+then
+    rm -f ~/.homebridge/accessories/cachedAccessories
+fi
 npm install -g .
 launchctl load ~/Library/LaunchAgents/com.homebridge.server.plist
 tail -f ~/.homebridge/homebridge.log
